Revert "Added NPC inventory."
This reverts commit 32f93d076b54eb2d2a8c86e362b4f6b1f90f56fe.
