+++ src/runtime/item/ItemFilterRuntime.ts
@@ -84,12 +84,6 @@ export class ItemFilterRuntime {
         defaultItemFilters.activationCostBonus.name,
         defaultItemFilters.activationCostReaction.name,
       ]),
-      [CONSTANTS.TAB_NPC_INVENTORY]: new Set<string>([
-        defaultItemFilters.activationCostAction.name,
-        defaultItemFilters.activationCostBonus.name,
-        defaultItemFilters.activationCostReaction.name,
-        defaultItemFilters.equipped.name,
-      ]),
     },
     [CONSTANTS.SHEET_TYPE_VEHICLE]: {
       [CONSTANTS.TAB_ACTOR_ACTIONS]: new Set<string>([
@@ -188,18 +182,6 @@ export class ItemFilterRuntime {
           defaultItemFilters.activationCostOther,
         ],
         ...getStandardSpellSchoolFilterCategories(),
-      },
-      [CONSTANTS.TAB_NPC_INVENTORY]: {
-        'DND5E.ItemActivationCost': [
-          defaultItemFilters.activationCostAction,
-          defaultItemFilters.activationCostBonus,
-          defaultItemFilters.activationCostReaction,
-        ],
-        'DND5E.Rarity': () => getItemRarityFilters(),
-        'TIDY5E.ItemFilters.Category.Miscellaneous': () => [
-          defaultItemFilters.equipped,
-          ...getAttunementFilters(),
-        ],
       },
       [CONSTANTS.TAB_ACTOR_ACTIONS]: { ...getActionListFilterCategories() },
     },
