+++ src/sheets/character/parts/InventoryGrid.svelte
@@ -6,15 +6,15 @@
     NpcSheetContext,
   } from 'src/types/types';
   import type { Item5e } from 'src/types/item.types';
-  import ItemTable from '../../components/item-list/v1/ItemTable.svelte';
-  import ItemTableHeaderRow from '../../components/item-list/v1/ItemTableHeaderRow.svelte';
-  import ItemTableColumn from '../../components/item-list/v1/ItemTableColumn.svelte';
+  import ItemTable from '../../../components/item-list/v1/ItemTable.svelte';
+  import ItemTableHeaderRow from '../../../components/item-list/v1/ItemTableHeaderRow.svelte';
+  import ItemTableColumn from '../../../components/item-list/v1/ItemTableColumn.svelte';
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import { CONSTANTS } from 'src/constants';
-  import GridPaneFavoriteIcon from '../../components/item-grid/GridPaneFavoriteIcon.svelte';
+  import GridPaneFavoriteIcon from '../../../components/item-grid/GridPaneFavoriteIcon.svelte';
   import { getContext } from 'svelte';
   import type { Readable, Writable } from 'svelte/store';
-  import TextInput from '../../components/inputs/TextInput.svelte';
+  import TextInput from '../../../components/inputs/TextInput.svelte';
   import { settingStore } from 'src/settings/settings';
   import { ActorItemRuntime } from 'src/runtime/ActorItemRuntime';
   import { declareLocation } from 'src/types/location-awareness.types';
