+++ src/sheets/character/parts/InventoryList.svelte
@@ -1,32 +1,31 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import type { Item5e } from 'src/types/item.types';
-  import ItemTable from '../../components/item-list/v1/ItemTable.svelte';
-  import ItemTableHeaderRow from '../../components/item-list/v1/ItemTableHeaderRow.svelte';
-  import ItemTableRow from '../../components/item-list/v1/ItemTableRow.svelte';
-  import ItemTableFooter from '../../components/item-list/ItemTableFooter.svelte';
-  import ItemTableColumn from '../../components/item-list/v1/ItemTableColumn.svelte';
-  import ItemTableCell from '../../components/item-list/v1/ItemTableCell.svelte';
-  import ItemUseButton from '../../components/item-list/ItemUseButton.svelte';
-  import ItemName from '../../components/item-list/ItemName.svelte';
+  import ItemTable from '../../../components/item-list/v1/ItemTable.svelte';
+  import ItemTableHeaderRow from '../../../components/item-list/v1/ItemTableHeaderRow.svelte';
+  import ItemTableRow from '../../../components/item-list/v1/ItemTableRow.svelte';
+  import ItemTableFooter from '../../../components/item-list/ItemTableFooter.svelte';
+  import ItemTableColumn from '../../../components/item-list/v1/ItemTableColumn.svelte';
+  import ItemTableCell from '../../../components/item-list/v1/ItemTableCell.svelte';
+  import ItemUseButton from '../../../components/item-list/ItemUseButton.svelte';
+  import ItemName from '../../../components/item-list/ItemName.svelte';
   import { CONSTANTS } from 'src/constants';
-  import ItemUses from '../../components/item-list/ItemUses.svelte';
-  import ItemAddUses from '../../components/item-list/ItemAddUses.svelte';
-  import ItemDeleteControl from '../../components/item-list/controls/ItemDeleteControl.svelte';
-  import ItemEditControl from '../../components/item-list/controls/ItemEditControl.svelte';
-  import EquipControl from '../../components/item-list/controls/EquipControl.svelte';
-  import AttuneControl from '../../components/item-list/controls/AttuneControl.svelte';
-  import InlineFavoriteIcon from '../../components/item-list/InlineFavoriteIcon.svelte';
-  import ItemFavoriteControl from '../../components/item-list/controls/ItemFavoriteControl.svelte';
+  import ItemUses from '../../../components/item-list/ItemUses.svelte';
+  import ItemAddUses from '../../../components/item-list/ItemAddUses.svelte';
+  import ItemDeleteControl from '../../../components/item-list/controls/ItemDeleteControl.svelte';
+  import ItemEditControl from '../../../components/item-list/controls/ItemEditControl.svelte';
+  import EquipControl from '../../../components/item-list/controls/EquipControl.svelte';
+  import AttuneControl from '../../../components/item-list/controls/AttuneControl.svelte';
+  import InlineFavoriteIcon from '../../../components/item-list/InlineFavoriteIcon.svelte';
+  import ItemFavoriteControl from '../../../components/item-list/controls/ItemFavoriteControl.svelte';
   import { getContext } from 'svelte';
   import type { Readable } from 'svelte/store';
   import type {
     CharacterSheetContext,
     InventorySection,
-    NpcSheetContext,
     RenderableClassicControl,
   } from 'src/types/types';
-  import AmmoSelector from './AmmoSelector.svelte';
+  import AmmoSelector from '../../actor/AmmoSelector.svelte';
   import { settingStore } from 'src/settings/settings';
   import ActionFilterOverrideControl from 'src/components/item-list/controls/ActionFilterOverrideControl.svelte';
   import { coalesce } from 'src/utils/formatting';
@@ -47,8 +46,7 @@
    */
   export let visibleItemIdSubset: Set<string> | null = null;
 
-  let context =
-    getContext<Readable<CharacterSheetContext | NpcSheetContext>>('context');
+  let context = getContext<Readable<CharacterSheetContext>>('context');
 
   const localize = FoundryAdapter.localize;
   const weightUnit = FoundryAdapter.getWeightUnit();
